pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "removebg"
version = "0.1.0"
description = "A simple CLI tool that removes backgrounds from images using rembg (U^2-Net)."
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [{ name = "removebg" }]
dependencies = [
    "click>=8.1",
    "rembg>=2.0.50",
    "Pillow>=10.0",
    "onnxruntime>=1.16",
]

[project.optional-dependencies]
gpu = ["onnxruntime-gpu>=1.16"]
dev = ["pytest>=7"]

[project.scripts]
removebg = "removebg.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"
